// src/api/generated/models/getOfficeHistoryApiV1OfficesOfficeIdHistoryGetParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */

export type GetOfficeHistoryApiV1OfficesOfficeIdHistoryGetParams = {
  /**
   * @minimum 1
   */
  page?: number
  /**
   * @minimum 1
   * @maximum 100
   */
  size?: number
  start_date?: string | null
  end_date?: string | null
}