// src/api/generated/models/paginatedResponseOfficeHistoryResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */
import type { OfficeHistoryResponse } from './officeHistoryResponse'

export interface PaginatedResponseOfficeHistoryResponse {
  data: OfficeHistoryResponse[]
  /** @minimum 0 */
  total: number
  /** @minimum 1 */
  page: number
  /** @minimum 1 */
  size: number
  /** @minimum 0 */
  pages: number
}